All test passing after merge
