+++ spec/models/card_spec.rb
@@ -118,7 +118,7 @@ describe Card do
     end
 
     it "should have updates" do
-      Wagn::Set::All::AttributeTracking::Model::Updates.should === @c.updates
+      Wagn::Set::All::AttributeTracking::Updates.should === @c.updates
     end
 
     it "should return original value" do
