+++ wagn-app/core-sets/all/pattern.rb
@@ -24,7 +24,7 @@ module ClassMethods
   end
 
   def find_set_model_module mod
-    module_name_parts = mod.split('::') << 'model'
+    module_name_parts = mod.split('::')
     module_name_parts.inject Wagn::Set do |base, part|
       return if base.nil?
       #Rails.logger.warn "find m #{base}, #{part}"
